Refactor Nostr context hooks
